backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
PyPDF2==3.0.1
google-generativeai==0.3.1
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.4
python-docx==0.8.11
pandas==2.1.0 